fix: 반려동물 등록시 몸무게를 Float > Double로 변경 (#150)

+++ src/main/java/com/kau/capstone/domain/pet/dto/request/PetRegistRequest.java
@@ -25,7 +25,7 @@ public class PetRegistRequest {
     @NotNull
     private Integer gender;
 
-    private float weight;
+    private Double weight;
 
     @NotNull
     private Integer age;
